requirements.txt
django
djangorestframework
django-cors-headers
feedparser
pytz